Fix undefined behavior in Vec::decode
This code is creating a Vec<u8> then transmuting it to Vec<T>.
Freeing or reallocating the underlying buffer then becomes undefined behavior
as the size and alignment of the buffer type has changed.

The fix here instead rewrites read_vec_u8 to create and resize
Vec<T>, only casting the underlying buffer to &[u8] when reading
from input.

There is no transmuting of Vec.

Found and and fix verified with miri.

Signed-off-by: Brian Anderson <andersrb@gmail.com>